+++ checkpoint/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 `CompositeCheckpointHandler.metadata()` to retrieve item metadata by
 default-constructing `CheckpointHandler`s when they're listed in the saved
 `StepMetadata` but aren't found in the checkpoint.
+- `FileOptions.format` to specify the underlying checkpointing file format.
 
 ### Fixed
 - Ignore not-exists and not-dir errors while building step metadata in
@@ -19,6 +20,7 @@ default-constructing `CheckpointHandler`s when they're listed in the saved
 
 ### Changed
 - Return `StepMetadata` from `CompositeCheckpointHandler.metadata()`.
+- `Checkpointer.save()` also saves `StepMetadata`.
 
 ## [0.10.2] - 2024-12-04